fix validation and default helpers
